Moved config.py to src folder for consistency
