+++ main.py
@@ -5,7 +5,7 @@ import discord
 from discord.ext import commands
 from discord_slash import SlashCommand
 
-from config import TOKEN
+from src.config import TOKEN
 
 bot = commands.Bot(command_prefix="!", intents=discord.Intents.all())
 slash = SlashCommand(bot, sync_commands=True)
